requirements.txt
chromadb==0.4.22
sentence-transformers==2.3.1
nltk==3.8.1
langchain==0.1.4
langchain-community==0.0.16
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1
openai==1.10.0
transformers==4.36.2
torch==2.1.2
peft==0.7.1
python-dotenv==1.0.0
